create shared section container component
