+++ webview/src/plots/components/styles.module.scss
@@ -8,38 +8,6 @@ $gap: 20px;
   height: 100%;
 }
 
-.plotsContainer {
-  width: 100%;
-
-  summary {
-    list-style-type: none;
-    font-family: var(--vscode-font-family);
-    margin: 14px 10px;
-    font-weight: bold;
-    font-size: 1.25rem;
-    display: flex;
-    align-items: center;
-    width: max-content;
-    cursor: pointer;
-  }
-}
-
-.plotsContainerWrapper {
-  position: relative;
-}
-
-.detailsIcon {
-  fill: $fg-color;
-  margin-bottom: -4px;
-}
-
-.iconMenu {
-  position: absolute;
-  right: $gap;
-  top: -15px;
-  z-index: 10;
-}
-
 .plotsWrapper {
   width: 100%;
 }
@@ -253,38 +221,6 @@ $gap: 20px;
     background-color: transparent !important;
     width: 100%;
     height: 100%;
-  }
-}
-
-.infoTooltipToggle {
-  display: flex;
-  align-items: center;
-  cursor: initial;
-}
-
-.infoIcon {
-  fill: $accent-color;
-  margin-left: 6px;
-}
-
-.infoTooltip {
-  max-width: 340px;
-  margin-left: 5px;
-  white-space: normal;
-  display: flex;
-  gap: 4px;
-  line-height: 16px;
-  font-size: 13px;
-  font-weight: 400;
-  cursor: auto;
-
-  svg {
-    min-width: 16px;
-    min-height: 16px;
-  }
-
-  a {
-    text-decoration: none;
   }
 }
